backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
audioop-lts==0.2.2
certifi==2025.11.12
cffi==2.0.0
click==8.3.1
colorama==0.4.6
cryptography==46.0.3
distro==1.9.0
ecdsa==0.19.1
fastapi==0.124.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.12.0
numpy==2.3.5
openai==2.8.1
pyasn1==0.6.1
pycparser==2.23
pydantic==2.12.4
pydantic_core==2.41.5
pyserial==3.5
python-jose==3.5.0
python-multipart==0.0.20
rsa==4.9.1
scipy==1.16.3
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.3
SpeechRecognition==3.14.4
standard-aifc==3.13.0
standard-chunk==3.13.0
starlette==0.50.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0
